backend/requirements.txt
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.9.0
certifi==2026.2.25
cffi==2.0.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.7
constantly==23.10.4
cryptography==46.0.7
daphne==4.1.2
Django==6.0.4
django-channels==0.7.0
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
google-auth==2.49.2
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
oauthlib==3.3.1
packaging==26.0
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
pyOpenSSL==26.0.0
redis==7.4.0
requests==2.33.1
requests-oauthlib==2.0.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
ujson==5.12.0
urllib3==2.6.3
zope.interface==8.3
